Split functions from auto_dark.py into auto_utils.py. Also created auto_ramp.py to control HV rampup. Needs more work.
